after heavy refactor of renderer to use events

+++ src/aggregate/anim.cpp
@@ -135,7 +135,7 @@ struct MyObject : IPositionEditorJob, IRenderJob, IAnimationJob
 	void pre_render( RenderParams & render_params ) 
 	{  }
 
-	void get_bounds( double *x1, double *y1, double *x2, double *y2 ) 
+	void get_bounds( int *x1, int *y1, int *x2, int *y2 ) 
 	{
 		bounding_rect_single( path, 0, x1, y1, x2, y2);
