+++ src/aggregate/raster.cpp
@@ -147,7 +147,7 @@ struct RenderJob : IRenderJob
 
 
 
-	void get_bounds( double *x1, double *y1, double *x2, double *y2 ) 
+	void get_bounds( int *x1, int *y1, int *x2, int *y2 ) 
 	{
 		
 		agg::path_storage	proj_limb ;
